import/export: support --url-prefix
To accomplish that, change all the relevant functions to accept a new
"base-url" argument which is just a map that must contain :protocol,
:host, :port, and optional :version and :prefix keys.  If the :version
isn't set, default it to :v4.

When creating URL strings for client/get, Use a "base-url -> URL -> URI
-> ASCII" trip to check for problems and perform escaping.

Adjust the jetty utils to manage a *base-url* rather than just a *port*,
so that tests can interact with the :prefix, etc.