Normalize the web-router-service prefix too
Without this change a [global] url-prefix setting with a relative path
like "foo" will prevent the server from responding on /foo.  That's
because add-web-routing-config (via the tk hook) was stashing the
unnormalized url-prefix in the web-router-service config before
process-config! had a chance to fix it.

In order to expose/test the original problem, make the same adjustments
to the config during testing that we do when running the service.