(PDB-1226) Defer fact_values value_hash constraint
When running with a large number of threads (80) and a large number of
requests (benchmark -n 20000 -r 10), it appears that the fact_values
insertions during start up (via add-facts!) end up in a significant
number of races.  Defer the constraint check to narrow the window of
opportunity for a collision.