(PDB-2075) Create temp databases for tests
Instead of just recycling a fixed set of databases (previously 2) for
tests, use a new temporary database each time.  In addition to being
safer, this provides optimization opportunities and removes a potential
barrier to the possible parallelization of testing.

See the changes to install_from_source.markdown for new required testing
configuration.