(PDB-2734) Ported MQ code to use the new threadpool ns
This change is fairly minor, however there is now a threadpool namespace
that contains similar logic to what is in mq-listener, except that the
new threadpool code is driven by a core.async channel. This commit moves
the current code over to the new threadpool code to prepare for a
cutover to stockpile.