(PDB-2734) Import/export fixes for stockpile
With the addition of Stockpile, the command API is changing a bit. Each
command needs to have a command type, version and certname to be
submitted along with a stream that represents the command. Currently
import/export works on strings and doesn't have the certname. This
commit extracts the certname from the exported filename and as a hack
around the stream problem, converts the string to bytes and an input stream.