(PDB-4271) query for fact expiration
Add two new fields to the node response: "expires_facts" and
"expires_facts_updated"

If "expires_facts" is true (the default), the node is subject to
garbage collection and expiration. If the value is set to false, the
nodes facts are excluded from the garbage collection procedure. If
this value has been explicitly set via a command, the
expires_facts_updated field will be a timestamp representing when the
field was set.